two-step: Add per mode settings
We want theme files to be able to specify different settings for
different modes ("boot-up" / "shutdown" / "updates"). Specifically we
want themes to be able to specify a text for (offline) updates mode to
tell the user what is going on, see the mockups at:
https://wiki.gnome.org/Design/OS/BootProgress

This commit adds support for per mode settings to the two-step plugins
and for starters moves the UseFirmwareBackground setting there, since we
don't want to show the firmware-background when showing the help-text.

Follow-up commits will add support for specifying the (optional) per mode
text to show, note eventually we will need to make these texts translatable.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>