ply-pixel-buffer: Fix bottom edge rendering of scaled buffers
Commit 4e1c00b89a71 (" ply-pixel-buffer: Fix right and bottom edge rendering
of scaled buffers"), which tried to fix a minor rendering issue with the right
and bottom edge of scaled images, contains a typo which actually makes things
worse :|

When checking if iy exceeds the height of the image, ix gets set to the height
instead of iy leading to addressing pass part of the buffer which leads to
various rendering artifacts.

This commit fixes the typo and thus also the artifacts.

Fixes: #83

Signed-off-by: Hans de Goede <hdegoede@redhat.com>