two-step: Change keyboard-indicator positioning to fixed offset below dialog
The keyboard-indicators should be located close to the dialog/entry to
make clear that they belong to each other. Using relative positioning
does not work well for this.

Drop the KeyboardIndicator[Vertic|Horizont]alAlignment options (we have
never done a release with these) and put the keyboard indicators at
a fixed offset (half their height) from the bottom of the dialog instead.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>