event-loop: Remove ply_trace calls around the disconnect handler
Doing a ply_trace both before and after the disconnect handler, which
gets called every time a boot client asks something of us through by
calling /sbin/plymouth leads to a lot of not really informative messages
in the debug-log.

This removes the 2 ply_trace calls around the disconnect handlers to make
the logs easier to read.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>