Fix Text fallback theme to match pop os themeing
The text fallback theme gets swapped to upon any rendering issues.
For example, when (un)plugging a monitor while displaying plymouth,
or toggling plymouth during a system upgrade.

This commit replaces the ubuntu-text theme with a pop-text theme,
so when users are in this error state their system doesn't look
like Ubuntu.